save loading and unloading projects

+++ OfferGUI/routes.py
@@ -9,8 +9,12 @@ from OfferGUI.tools import AddRow, DelRow, XmlReader, RemoveTemporaryItems
 import os
 import xmltodict
 @app.route('/costs', methods=['POST', 'GET'])
-# @login_required
+@login_required
 def costs_page():
+    print("yeah")
+    if len(temp_inquiry.query.all()) == 0 and len(temp_project_info.query.all()) == 0:
+       flash(f"No project selected!", category='info')
+       return redirect(url_for('home_page'))  
     # if len(temp_project_info.query.all()) == 0:
     #     flash(f"No project selected!", category='info')
     #     return redirect(url_for('home_page'))  
@@ -113,6 +117,7 @@ def register_page():
 
 @app.route('/login', methods=['GET', 'POST'])
 def login_page():
+    RemoveTemporaryItems()
     form = LoginForm()
     if form.validate_on_submit():
         attempted_user = User.query.filter_by(username=form.username.data).first()
@@ -128,81 +133,45 @@ def login_page():
 
 @app.route('/logout', methods=['GET', 'POST'])
 def logout_page():
+    RemoveTemporaryItems()
     logout_user()
     flash("You have been logged out!", category='info')
     return redirect(url_for("home_page"))
 
 @app.route('/', methods=['GET', 'POST'])
 @app.route('/home', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def home_page():
-    save_form = SaveForm()
     home_form = HomeForm()
-    form = ProjectForm()
     db_projects = collected_projects.query.with_entities(collected_projects.project_name, collected_projects.project_name).filter(collected_projects.project_name!="NULL")
     home_form.project.choices = [k for k in db_projects]
     if request.method == 'POST':
         if request.form.get('LoadExistingProject'):
             RemoveTemporaryItems()
+            print("removed!")
             req_project_name = [request.form[key] for key in request.form.keys()][1]
-            print(req_project_name)
+            ### get row in collected projects by select field above 'load'
             project_row = collected_projects.query.filter_by(project_name=req_project_name).first()
             print(project_row.id)
-            ### class from from models-py
-            dde = dropdown_elements
-            ### reading from database
-            db_user = User.query.with_entities(User.username, User.username).all()
-            db_plant_type = dde.query.with_entities(dde.plant_type, dde.plant_type).filter(dde.plant_type!="NULL")
-            db_busbar = dde.query.with_entities(dde.busbar, dde.busbar).filter(dde.busbar!="NULL")
-            db_calc_for = dde.query.with_entities(dde.calc_for, dde.calc_for).filter(dde.calc_for!="NULL")
-            db_yes_no = dde.query.with_entities(dde.yes_no, dde.yes_no).filter(dde.yes_no!="NULL")
-            db_languages = dde.query.with_entities(dde.languages, dde.languages).filter(dde.languages!="NULL")
-            db_temp_inquiry = temp_inquiry.query.all()
-            ### setting choices (forms.py --> class ProjectForm)
-            form.calc_for.choices = [k for k in db_calc_for]
-            form.editor.choices = [k for k in db_user]
-            form.plant_type.choices = [k for k in db_plant_type]
-            form.busbar.choices = [k for k in db_busbar]
-            form.commissioning.choices = [k for k in db_yes_no]
-            form.site_management.choices = [k for k in db_yes_no]
-            form.manpower.choices = [k for k in db_yes_no]
-            form.manpower_language.choices = [k for k in db_languages]
-            form.tools.choices = [k for k in db_yes_no]
-            ### get temp inquiry info from database
-            item = project_row
 
-            ### set data on project info page
-            form.editor.data                = item.editor
-            form.project_name.data          = item.project_name
-            form.project_manager_dept.data  = item.project_manager_dept
-            form.site.data                  = item.site
-            form.calc_for.data              = item.calc_for
-            form.number_of_bays.data        = item.number_of_bays
-            form.plant_type.data            = item.plant_type
-            form.busbar.data                = item.busbar
-            form.date.data                  = datetime.strptime(item.date, '%Y-%m-%d')
-            form.cost_determination.data    = datetime.strptime(item.cost_determination, '%Y-%m-%d')
-            # form.commissioning.data         = item.commissioning
-            # form.manpower.data              = item.manpower
-            # form.manpower_language.data     = item.manpower_language
-            # form.tools.data                 = item.tools
             ### writing in database
-            temp_project_infos = temp_project_info(project_name         = item.project_name        ,
-                                                   project_manager_dept = item.project_manager_dept,
-                                                   order_indicator      = item.order_indicator     ,
-                                                   site                 = item.site                ,
-                                                   customer             = item.customer            ,
-                                                   calc_for             = item.calc_for            ,
-                                                   date                 = item.date                ,
-                                                   cost_determination   = item.cost_determination  ,
-                                                   editor               = item.editor              ,
-                                                   project_id           = item.project_id          ,
-                                                   plant_type           = item.plant_type          ,
-                                                   busbar               = item.busbar              ,
-                                                   number_of_bays       = item.number_of_bays      )
+            temp_project_infos = temp_project_info(project_name         = project_row.project_name        ,
+                                                   project_manager_dept = project_row.project_manager_dept,
+                                                   order_indicator      = project_row.order_indicator     ,
+                                                   site                 = project_row.site                ,
+                                                   customer             = project_row.customer            ,
+                                                   calc_for             = project_row.calc_for            ,
+                                                   date                 = project_row.date                ,
+                                                   cost_determination   = project_row.cost_determination  ,
+                                                   editor               = project_row.editor              ,
+                                                   project_id           = project_row.project_id          ,
+                                                   plant_type           = project_row.plant_type          ,
+                                                   busbar               = project_row.busbar              ,
+                                                   number_of_bays       = project_row.number_of_bays      ,
+                                                   commissioning        = project_row.commissioning       )
             db.session.add(temp_project_infos)
             db.session.commit()
-            return render_template('project_info.html', form=form, save_form=save_form)
+            return redirect(url_for('project_page'))
     return render_template('home.html', home_form=home_form)
 
 @app.route('/uploader', methods = ['POST'])
@@ -221,7 +190,7 @@ def upload_file():
             return redirect(url_for('home_page'))
 
 @app.route('/project', methods=['GET', 'POST'])
-# @login_required
+@login_required
 def project_page():
     save_form = SaveForm()
     form = ProjectForm()
@@ -236,6 +205,7 @@ def project_page():
         db_yes_no = dde.query.with_entities(dde.yes_no, dde.yes_no).filter(dde.yes_no!="NULL")
         db_languages = dde.query.with_entities(dde.languages, dde.languages).filter(dde.languages!="NULL")
         db_temp_inquiry = temp_inquiry.query.all()
+        temp_project_info.query.all()
         # setting choices (forms.py --> class ProjectForm)
         form.calc_for.choices = [k for k in db_calc_for]
         form.editor.choices = [k for k in db_user]
@@ -253,33 +223,38 @@ def project_page():
             form.editor.data                = item.editor
             form.project_name.data          = item.project_name
             form.project_manager_dept.data  = item.project_manager_dept
+            form.order_indicator.data       = item.order_indicator
             form.site.data                  = item.site
+            form.customer.data              = item.customer
             form.calc_for.data              = item.calc_for
+            form.date.data                  = datetime.strptime(item.date, '%Y-%m-%d')
+            form.cost_determination.data    = datetime.strptime(item.cost_determination, '%Y-%m-%d')
+            form.editor.data                = item.editor
+            form.project_id.data            = item.project_id
             form.number_of_bays.data        = item.number_of_bays
             form.plant_type.data            = item.plant_type
             form.busbar.data                = item.busbar
-            form.date.data                  = datetime.strptime(item.date, '%Y-%m-%d')
-            form.cost_determination.data    = datetime.strptime(item.cost_determination, '%Y-%m-%d')
-            # form.commissioning.data         = item.commissioning
+            form.commissioning.data         = item.commissioning
             # form.manpower.data              = item.manpower
             # form.manpower_language.data     = item.manpower_language
             # form.tools.data                 = item.tools      
               
         # setting data by imported xml-file (forms.py --> class ProjectForm)
-        if len(temp_inquiry.query.all()) == 1:
+        elif len(temp_inquiry.query.all()) == 1:
             #get temp inquiry info from database
             item = temp_inquiry.query.get(1)
             #set data on project info page
             form.editor.data                = current_user.username
             form.project_name.data          = item.project_name
             form.project_manager_dept.data  = item.firstname + " " + item.lastname + " / " + item.department
-            form.site.data                  = item.city + ", " + item.country
+            form.site.data                  = str(item.city) + ", " + str(item.country)
             form.calc_for.data              = item.calc_for
             form.number_of_bays.data        = item.number_of_bays
             form.plant_type.data            = item.plant_type
             form.busbar.data                = item.busbar
             form.date.data                  = datetime.today()
-            form.cost_determination.data    = datetime.strptime(item.offer_until, '%Y-%m-%d')
+            if item.offer_until != None:
+                form.cost_determination.data    = datetime.strptime(item.offer_until, '%Y-%m-%d')
             form.commissioning.data         = item.commissioning
             form.manpower.data              = item.mpd
             form.manpower_language.data     = item.language
@@ -299,44 +274,58 @@ def project_page():
         #                             and attr not in ['Save project!','metadata','query','query_class']]
         project_output = [request.form[key] for key in request.form.keys()]
         # print(project_output)
+        if len(project_output) >= 1:
+            # writing in database
+            project_infos = collected_projects(project_name         = project_output[2],
+                                            project_manager_dept = project_output[3],
+                                            order_indicator      = project_output[4],
+                                            site                 = project_output[5],
+                                            customer             = project_output[6],
+                                            calc_for             = project_output[7],
+                                            date                 = project_output[8],
+                                            cost_determination   = project_output[9],
+                                            editor               = project_output[10],
+                                            project_id           = project_output[11],
+                                            plant_type           = project_output[12],
+                                            busbar               = project_output[13],
+                                            number_of_bays       = project_output[14])
+            #temp data needed for cost page
+            # temp_project_infos = temp_project_info(project_name=project_output[2],
+            #                                        calc_for=project_output[7],
+            #                                        editor=project_output[10],
+            #                                        project_id=project_output[11],)
+            temp_project_infos = temp_project_info(project_name         = project_output[2],
+                                                project_manager_dept = project_output[3],
+                                                order_indicator      = project_output[4],
+                                                site                 = project_output[5],
+                                                customer             = project_output[6],
+                                                calc_for             = project_output[7],
+                                                date                 = project_output[8],
+                                                cost_determination   = project_output[9],
+                                                editor               = project_output[10],
+                                                project_id           = project_output[11],
+                                                plant_type           = project_output[12],
+                                                busbar               = project_output[13],
+                                                number_of_bays       = project_output[14]) 
+            temp_project_info.query.delete()
+            db.session.commit()
+            db.session.add(project_infos)
+            db.session.add(temp_project_infos)
+            db.session.commit()
+            flash(f"Data from project page stored in database!", category='success')
 
-        # writing in database
-        project_infos = collected_projects(project_name         = project_output[2],
-                                           project_manager_dept = project_output[3],
-                                           order_indicator      = project_output[4],
-                                           site                 = project_output[5],
-                                           customer             = project_output[6],
-                                           calc_for             = project_output[7],
-                                           date                 = project_output[8],
-                                           cost_determination   = project_output[9],
-                                           editor               = project_output[10],
-                                           project_id           = project_output[11],
-                                           plant_type           = project_output[12],
-                                           busbar               = project_output[13],
-                                           number_of_bays       = project_output[14])
-        #temp data needed for cost page
-        temp_project_infos = temp_project_info(project_name=project_output[2],
-                                               calc_for=project_output[7],
-                                               editor=project_output[10],
-                                               project_id=project_output[11],)
-        temp_project_info.query.delete()
-        db.session.add(project_infos)
-        db.session.add(temp_project_infos)
-        db.session.commit()
-        flash(f"Data from project page stored in database!", category='success')
 
+            # project_output = [[request.form[key] for key in request.form.keys() if key == i] for i in project_info_col]
+            # print(project_output)
 
-        # project_output = [[request.form[key] for key in request.form.keys() if key == i] for i in project_info_col]
-        # print(project_output)
+            # for attr, value in project_info.__dict__.items():
+            #     attr = project_info_col[i]
+            #     value = project_output[i]
 
-        # for attr, value in project_info.__dict__.items():
-        #     attr = project_info_col[i]
-        #     value = project_output[i]
+            # for i,j in zip(project_info_col, project_output):
+            #     project_infos = project_info.__dict__.keys
 
-        # for i,j in zip(project_info_col, project_output):
-        #     project_infos = project_info.__dict__.keys
-
-        return redirect(url_for('costs_page'))
+            return redirect(url_for('costs_page'))
     if save_form.errors != {}: #If there are not errors from the validations
         for err_msg in form.errors.values():
             flash(f'There was an error with creating a user: {err_msg}', category='danger')
